syndecrypt/requirements.txt
pycrypto == 2.6.1
docopt == 0.6.2

assertpy == 0.9
pytest == 2.9.2

coverage == 4.0.3
codacy-coverage == 1.3.2
